test: one definition of what a synthetic product looks like
